+++ Zal_User/Controller/Tabbar/Home/NearStoreCell.swift
@@ -36,11 +36,11 @@ class NearStoreCell: UICollectionViewCell,NearStoreCellView {
         cocntentView.layer.applySketchShadow()
     }
 
-    func Data(name: String, ProfileImage: String, address: String, CoverImage: String, rate: Int, distance: Double) {
+    func Data(name: String, ProfileImage: String, address: String, CoverImage: String, rate: Double, distance: Double) {
         nameLabel.text = name
         addressLabel.text = address
         distanceLabel.text = String(distance)
-        rateView.rating = Double(rate)
+        rateView.rating = rate
         let url = URL(string: ProfileImage)
         profileImage.kf.setImage(with: url)
         let url2 = URL(string: CoverImage)
